Refactor, add tree folder
