+++ src/extension.ts
@@ -2,7 +2,6 @@
 // Import the module and reference it with the alias vscode in your code below
 import axios from 'axios';
 import * as apiClient from '@hackmd/api';
-import { MdTreeItemProvider } from './mdTreeView';
 import { MdTextDocumentContentProvider } from './mdTextDocument';
 import * as vscode from 'vscode';
 import * as markdownitContainer from 'markdown-it-container';
@@ -10,6 +9,8 @@ import * as S from 'string';
 import { store } from './store'
 import { initializeStorage } from './store/storage'
 import * as Prism from 'prismjs';
+import { checkLogin, login, refreshHistoryList } from './tree/index';
+import { HackMDTreeViewProvider } from './tree/index'
 
 require('prismjs/components/prism-wiki');
 require('prismjs/components/prism-haskell');
@@ -230,43 +231,6 @@ let highlight;
 const API = new apiClient.default();
 axios.defaults.withCredentials = true;
 
-
-const refreshHistoryList = async (context) => {
-  if (await checkLogin()) {
-    store.history = (await API.getHistory()).history;
-    context.globalState.update('history', store.history);
-  } else {
-    store.history = [{}];
-    context.globalState.update('history', [{}]);
-  }
-};
-
-const checkLogin = async () => {
-  return (await API.getMe()).status === 'ok';
-};
-
-const login = async (context: vscode.ExtensionContext) => {
-  const { email, password } = getLoginCredential(context);
-  if (!email || !password) {
-    vscode.window.showInformationMessage('Please enter your email and password to use HackMD extension!')
-    return;
-  }
-  await API.login(email, password);
-  if (await checkLogin()) {
-    store.isLogin = true;
-    context.globalState.update('isLogin', true);
-    vscode.window.showInformationMessage('Successfully login!');
-  } else {
-    vscode.window.showInformationMessage('Wrong email or password, please enter again');
-  }
-};
-
-const getLoginCredential = (context: vscode.ExtensionContext) => {
-  const email: string = context.globalState.get('email');
-  const password: string = context.globalState.get('password');
-  return { email, password };
-};
-
 export async function activate(context: vscode.ExtensionContext) {
   initializeStorage(context);
   context.subscriptions.push(vscode.commands.registerCommand('HackMD.login', async () => {
@@ -304,7 +268,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
     context.globalState.update('email', email);
     context.globalState.update('password', password);
-    
+
     await login(context);
     await refreshHistoryList(context);
   }));
@@ -321,12 +285,12 @@ export async function activate(context: vscode.ExtensionContext) {
     await refreshHistoryList(context);
   }));
 
-  const treeViewProvider = new MdTreeItemProvider(store);
+  const hackMDTreeViewProvider = new HackMDTreeViewProvider(store);
   context.subscriptions.push(
-    vscode.window.registerTreeDataProvider('mdTreeItems', treeViewProvider)
+    vscode.window.registerTreeDataProvider('mdTreeItems', hackMDTreeViewProvider)
   );
   context.subscriptions.push(
-    vscode.commands.registerCommand('treeView.refreshList', () => treeViewProvider.refresh())
+    vscode.commands.registerCommand('treeView.refreshList', () => hackMDTreeViewProvider.refresh())
   );
 
   context.subscriptions.push(vscode.workspace.registerTextDocumentContentProvider('hackmd', new MdTextDocumentContentProvider()));
